backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
requests==2.31.0
transformers==4.45.0
torch==2.4.1